Share SO3 quaternion helpers
